refactor(automation): name repeated literals for the goconst gate
The strict lint gate flags repeated string literals across the new
automation files. Two moves, no behaviour change:

- Extract one shared taskCreateEffect(ev, subject, dueAt) builder so the
  create_task effect's JSON keys (task/subject/due_at/links/entity_type/
  entity_id) live in one place instead of three hand-copied maps.
  stageChangeCreateTask and route_lead now delegate through it; the clock
  handlers' anchorReminderTaskEffect does too. Byte-identical: the one
  literal that changed form (stage_change's entity_type "deal" ->
  string(ev.Entity.Type)) resolves to "deal" for its deal.stage_changed
  trigger.
- Name the JSON-schema vocabulary (object/string/properties/
  additionalProperties/description), the routing-rule keys, and the
  shared "not a parameter" 422 reason as consts.

The const additions pushed automations_catalog.go past the hard 500-LOC
cap, so the self-contained lead-routing params schema and its validators
move to catalog_leadrouting_schema.go (a clean concept split, not a
waiver).
